+++ src/routes/thread/comment/index.ts
@@ -6,6 +6,9 @@ import unpin from "./unpin";
 import vote from "./vote";
 import images from "./images";
 import emotion from "./emotion";
+import emotions from "./emotions";
+import votes from "./votes";
+import checkRemoved from "../../../plugins/checkComment";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 
 export default function (
@@ -13,6 +16,7 @@ export default function (
     _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) {
+    fastify.addHook("preHandler", checkRemoved);
     fastify.register(comment);
     fastify.register(replies);
     fastify.register(create);
@@ -20,6 +24,8 @@ export default function (
     fastify.register(unpin);
     fastify.register(vote);
     fastify.register(images);
-    fastify.register(emotion);
+    fastify.register(emotions);
+    fastify.register(votes);
+    fastify.register(emotion, { prefix: "/:cid/emotion" });
     done();
 }