+++ src/routes/thread/comment/pin.ts
@@ -63,7 +63,6 @@ export default function (
             ) as commentType;
 
             if (!comment) return res.code(404).send({ error: "Comment not found." });
-            if (comment.removed) return res.code(410).send({ error: "Comment removed." });
 
             await threadCl.updateOne({ id: threadId }, { $set: { pin: comment } });