+++ src/routes/thread/comment/replies.ts
@@ -41,9 +41,6 @@ export default (
             if (!targetComment)
                 return res.code(404).send({ error: "Thread or comment not found." });
 
-            if (targetComment.removed)
-                return res.code(410).send({ error: "Comment removed." });
-
             const replies = (
                 await threadCl.findOne(
                     { id },