+++ src/routes/thread/index.ts
@@ -5,7 +5,7 @@ import thread from "./thread";
 import create from "./create";
 import comment from "./comment";
 import category from "./category";
-import hidden from "../../plugins/hidden";
+import checkHidden from "../../plugins/checkHidden";
 import star from "./actions/star";
 import unstar from "./actions/unstar";
 
@@ -14,7 +14,7 @@ export default (
     _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) => {
-    fastify.addHook("preHandler", hidden);
+    fastify.addHook("preHandler", checkHidden);
     fastify.register(thread);
     fastify.register(checkExist);
     fastify.register(images);